requirements.txt
numpy==2.1.2
pandas==2.2.3
pillow==11.0.0
python-dateutil==2.9.0.post0
pytz==2024.2
six==1.16.0
tzdata==2024.2